requirements.txt
python-dotenv==1.0.1
yfinance==0.2.40
Flask==3.0.3
Flask-WTF==1.2.1